created constants.js
